requirements.txt
flask==1.1.2
flask-socketio==4.3.0
PyJWT==1.7.1
pydantic==1.5.1
gunicorn==20.0.4
SQLAlchemy==1.3.17
redis==3.2.1
eventlet==0.25.2
Flask-Cors==3.0.8